// src/types/reportly_types.ts
export interface reportData {
  name?: string | null;
  phone_Number?: string | null;
  wallet_address?: string | null;
  fraudster_wallet_address?: string | null;
  description?: string | null;
  complaint?: string | null;
  status?: string | null;
  report_id?: string | null;
  confirmer?: string | null;
}